Ensure we receive messages at the top layer of the stack
This means we need to track what messages are being produced at the
_base_ of the stack (ProcessTM userMsg handledMsg) by the time it
is being run/spawned. Thus things like (spawn (lift receive)) will
not typecheck.

Note that handledMsg is not the type of messages received on the
channel! Those are some unknown foreign type. But any nontrivial
MonadProcessTrans instance will change the outMsg type, so it is
safe.